+++ extensions/amp-auto-ads/0.1/adsense-network-config.js
