+++ extensions/amp-auto-ads/0.1/denakop-network-config.js
