+++ extensions/amp-auto-ads/0.1/doubleclick-network-config.js
