+++ extensions/amp-auto-ads/0.1/ping-network-config.js
